backend/requirements.txt
asgiref==3.7.2
astroid==3.0.1
black==24.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.7
distlib==0.3.7
Django==3.2.16
django-ckeditor==6.3.2
django-cors-headers==3.13.0
django-environ==0.9.0
django-js-asset==2.2.0
django-storages==1.13.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
filelock==3.13.1
idna==3.6
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
pi==0.1.2
Pillow==9.3.0
platformdirs==4.0.0
psycopg2==2.9.5
PyJWT==2.8.0
pylint==3.0.2
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tomlkit==0.12.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
virtualenv==20.25.0
